requirements.txt
Flask==2.3.3
librosa==0.10.1
SpeechRecognition==3.10.0
Werkzeug==2.3.7
numpy==1.25.2
scipy==1.11.2
openai-whisper
ffmpeg-python
